pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bovin_demo"
version = "0.1.0"
description = "BOVIN-Pathway HeteroGNN demo — Aim 1 minimum closed loop (TCGA-COAD → ICD-readiness + module XAI)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Nabe", email = "z4fu@ucsd.edu" }]
keywords = ["bioinformatics", "gnn", "icd", "immunotherapy", "xai"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]

# ============== Runtime dependencies ==============
# NOTE: torch / torch-geometric 在 Docker 里单独 pin-install（见 Dockerfile），
# 不在 pyproject 直接锁版本 —— 避免 pip resolver 在 Apple Silicon 上爆炸。
dependencies = [
  "numpy>=1.26,<2.0",
  "pandas>=2.1,<2.3",
  "scipy>=1.11",
  "scikit-learn>=1.4,<1.6",
  "pyyaml>=6.0",
  "omegaconf>=2.3",
  "hydra-core>=1.3,<1.4",
  "pydantic>=2.5,<3.0",
  "pytorch-lightning==2.2.*",
  "captum>=0.7",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "tqdm>=4.66",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.4.0",
  "pre-commit>=3.6",
  "mypy>=1.8",
  "ipykernel>=6.29",
  "jupyter>=1.0",
  "networkx>=3.2",  # T1.6 graph overview PNG only
]

[project.scripts]
bovin-demo = "bovin_demo.cli:main"

[project.urls]
Repository = "https://github.com/<TODO>/bovin-pathway-demo"

# ============== setuptools ==============
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["bovin_demo*"]
exclude = ["tests*", "notebooks*"]

[tool.setuptools.package-data]
# M1 · T1.3: ship the pre-parsed pathway JSON so ``load_graph()`` has a
# zero-config default that works even when the source markdown isn't present.
"bovin_demo.graph" = ["*.json"]

# ============== Ruff ==============
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks", "outputs", "data"]

[tool.ruff.lint]
select = [
  "E", "W",   # pycodestyle
  "F",        # pyflakes
  "I",        # isort
  "B",        # flake8-bugbear
  "UP",       # pyupgrade
  "SIM",      # simplify
  "RUF",
]
ignore = [
  "E501",     # line-too-long — 让 formatter 处理
  "B008",     # function default argument — hydra / typer 常用
  "RUF001",   # ambiguous unicode in identifiers
  "RUF002",   # ambiguous unicode in docstrings — 我们故意用 × · – 等学术符号
  "RUF003",   # ambiguous unicode in comments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "SIM"]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============== Pytest ==============
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--tb=short",
]
markers = [
  "slow: deselect with '-m \"not slow\"'",
  "gpu: requires GPU (deselect by default in CI)",
]

# ============== Coverage ==============
[tool.coverage.run]
source = ["bovin_demo"]
branch = true
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]

# ============== Mypy ==============
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
